update check_past_kv logic
